+++ Clase Springboot/src/main/java/com/uade/tpo/demo/entity/dto/ProductoRequest.java
@@ -1,5 +1,5 @@
 package com.uade.tpo.demo.entity.dto;
-
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -9,8 +9,8 @@ public class ProductoRequest {
     private Long categoriaId;
     private String nombre;
     private String descripcion;
-    private Float precio;
+    private BigDecimal precio;
     private Integer stock;
-    private Float descuento;
+    private BigDecimal descuento;
     private List<String> imagenes;
 }
